requirements.txt

requests>=2.28.0
python-dotenv>=0.21.0
pandas>=1.5.0
openpyxl>=3.0.0  # Для работы с Excel